bhukkad/requirement.txt
Django==1.8.4
argparse==1.2.1
django-autoslug==1.8.0
wsgiref==0.1.2
